+++ src/main/arguments/arguments.py
@@ -202,7 +202,9 @@ class TuneArguments(TunerFunctionArguments):
                  mlm_probability: float = 0.15,
                  use_flash_attention: bool = False,
                  flash_attention_impl: str = 'flash_attention_2',
-                 push_adapter: bool = True
+                 push_adapter: bool = True,
+                 use_gradient_checkpointing: bool = True,
+                 overwrite_repo: bool = False
                  ):
         super(TuneArguments, self).__init__(new_model, is_fp16, is_bf16, use_4bit, use_8bit, fp32_cpu_offload, is_chat_model, padding_side, use_agent_tokens, additional_vocabulary_tokens, huggingface_auth_token, is_debug_mode=is_debug_mode)
         self.r = r
@@ -251,6 +253,8 @@ class TuneArguments(TunerFunctionArguments):
         self.use_flash_attention = use_flash_attention
         self.flash_attention_impl = flash_attention_impl
         self.push_adapter = push_adapter
+        self.use_gradient_checkpointing = use_gradient_checkpointing
+        self.overwrite_repo = overwrite_repo
 
     def validate(self) -> None:
         # I know it's bad, I will clean it up eventually
@@ -364,10 +368,14 @@ class PushArguments(TunerFunctionArguments):
                  use_agent_tokens: bool = False,
                  additional_vocabulary_tokens: list | None = None,
                  huggingface_auth_token: str | None = None,
-                 is_debug_mode: bool = False):
+                 is_debug_mode: bool = False,
+                 overwrite_repo: bool = False,
+                 target_repo: str | None = None):
         super(PushArguments, self).__init__(new_model, is_fp16, is_bf16, use_4bit, use_8bit, fp32_cpu_offload=fp32_cpu_offload, is_chat_model=is_chat_model, padding_side=padding_side, use_agent_tokens=use_agent_tokens, additional_vocabulary_tokens=additional_vocabulary_tokens, huggingface_auth_token=huggingface_auth_token, is_debug_mode=is_debug_mode)
         self.model_dir = model_dir
         self.public_push = public_push
+        self.overwrite_repo = overwrite_repo
+        self.target_repo = target_repo
 
     def validate(self) -> None:
         is_valid = self.new_model is not None and self.model_dir is not None